Move OCI SSE test helpers
Keep shared SSE-C assertions with SSE header coverage to avoid duplicating helper code across OCI storage tests.
